Changed most of the files.
The index.html has a script at the end that does the heavy lifting for converting the original BibTeX file to JSON. For now I am using jQuery.

As for the original file it takes the script approximately 4-5 seconds to convert it.

Now most of the files are a copy from the repository of https://github.com/mayanklahiri/bib2json